refactor: Advanced settings component
