+++ src/routes/safe/store/models/safe.ts
@@ -13,7 +13,7 @@ export type SafeRecordProps = {
   threshold: number
   ethBalance: string
   owners: List<{ name: string; address: string }>
-  modules: List<ModulePair> | null
+  modules: ModulePair[] | null
   activeTokens: Set<string>
   activeAssets: Set<string>
   blacklistedTokens: Set<string>
@@ -33,7 +33,7 @@ const makeSafe = Record<SafeRecordProps>({
   threshold: 0,
   ethBalance: '0',
   owners: List([]),
-  modules: List(),
+  modules: [],
   activeTokens: Set(),
   activeAssets: Set(),
   blacklistedTokens: Set(),
